+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 - Crash when rendering a node with a null material
 
+### Removed
+
+- Undocumented static helpers on `Renderable`
+
 ## [0.1.0] - 2026-09-01
 
 ### Added